Almost API-complete

+++ api/model_temperature_target.go
@@ -2,5 +2,5 @@ package api
 
 // TemperatureTarget - The target temperature of the boiler
 type TemperatureTarget struct {
-	Target float32 `json:"target,omitempty"`
+	Target float64 `json:"target,omitempty"`
 }
